Extract all operator tests int their dedicated class
