+++ core/common/src/test/kotlin/com/github/jcornaz/miop/operator/CombineLatestTest.kt
@@ -1,6 +1,8 @@
 package com.github.jcornaz.miop.operator
 
-import com.github.jcornaz.miop.*
+import com.github.jcornaz.miop.combineLatestWith
+import com.github.jcornaz.miop.failedReceiveChannel
+import com.github.jcornaz.miop.receiveChannelOf
 import com.github.jcornaz.miop.test.assertThrows
 import com.github.jcornaz.miop.test.runTest
 import kotlinx.coroutines.Dispatchers
@@ -12,58 +14,53 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
-class MergeTest : OperatorTest() {
-    override fun <T> ReceiveChannel<T>.identityOperation(): ReceiveChannel<T> =
-        mergeWith(emptyReceiveChannel())
+class CombineLatestTest : OperatorTest() {
 
+    override fun <T> ReceiveChannel<T>.identityOperation(): ReceiveChannel<T> =
+        combineLatestWith(receiveChannelOf(1)) { it, _ -> it }
 
     @Test
-    fun mergeShouldGiveTheElementsAsSoonAsReceived() = runTest {
+    fun combineLatestShouldGiveTheCombinedElements() = runTest {
         val source1 = Channel<Int>()
         val source2 = Channel<Int>()
-        val result = source1.mergeWith(source2)
+        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
 
         expect(1)
         launch(Dispatchers.Unconfined) {
             expect(2)
-            assertEquals(1, result.receive())
-            expect(4)
-            assertEquals(2, result.receive())
-            expect(6)
-            assertEquals(3, result.receive())
-            expect(8)
-            assertEquals(4, result.receive())
-            expect(10)
-            assertEquals(5, result.receive())
-            expect(12)
-            assertEquals(6, result.receive())
-            expect(14)
+            assertEquals(2 to 3, result.receive())
+            expect(5)
+            assertEquals(2 to 4, result.receive())
+            expect(7)
+            assertEquals(5 to 4, result.receive())
+            expect(9)
+            assertEquals(5 to 6, result.receive())
+            expect(11)
             assertThrows<ClosedReceiveChannelException> { result.receive() }
-            expect(16)
+            expect(13)
         }
         expect(3)
         source1.send(1)
-        expect(5)
         source1.send(2)
-        expect(7)
+        expect(4)
         source2.send(3)
-        expect(9)
+        expect(6)
         source2.send(4)
-        expect(11)
+        expect(8)
         source1.send(5)
         source1.close()
-        expect(13)
+        expect(10)
         source2.send(6)
-        expect(15)
+        expect(12)
         source2.close()
-        finish(17)
+        finish(14)
     }
 
     @Test
-    fun cancellingAChannelMergedByMergeWithShouldCancelAllTheSources() = runTest {
+    fun cancellingAChannelMergedByCombineLatestShouldCancelAllTheSources() = runTest {
         val source1 = Channel<Int>()
         val source2 = Channel<Int>()
-        val result = source1.mergeWith(source2)
+        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
 
         result.cancel()
 
@@ -75,10 +72,10 @@ class MergeTest : OperatorTest() {
     }
 
     @Test
-    fun ifASourceMergedByMergeWithFailsTheResultShouldFailAndTheOtherSourcesShouldBeCancelled() = runTest {
+    fun ifASourceMergedByCombineLatestFailsTheResultShouldFailAndTheOtherSourcesShouldBeCancelled() = runTest {
         val source1 = Channel<Int>()
         val source2 = failedReceiveChannel<Int>(Exception("my exception"))
-        val result = source1.mergeWith(source2)
+        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
 
         assertTrue(source1.isClosedForReceive, "other source should be closed for receive")
         assertTrue(result.isClosedForReceive, "result should be closed for receive")
@@ -88,11 +85,10 @@ class MergeTest : OperatorTest() {
     }
 
     @Test
-    fun itShouldBePossibleToUseMergeWithFiniteChannel() = runTest {
-        val result = receiveChannelOf(1).mergeWith(receiveChannelOf(2))
+    fun itShouldBePossibleToUseCombineLatestWithFiniteChannel() = runTest {
+        val result = receiveChannelOf(1).combineLatestWith(receiveChannelOf('a')) { v1, v2 -> v1 to v2 }
 
-        assertEquals(1, result.receive())
-        assertEquals(2, result.receive())
+        assertEquals(1 to 'a', result.receive())
         assertThrows<ClosedReceiveChannelException> { result.receive() }
     }
 }
