--- core/common/src/test/kotlin/com/github/jcornaz/miop/operator/OperatorsTest.kt
@@ -1,176 +1,0 @@
-package com.github.jcornaz.miop.operator
-
-import com.github.jcornaz.miop.*
-import com.github.jcornaz.miop.test.AsyncTest
-import com.github.jcornaz.miop.test.assertThrows
-import com.github.jcornaz.miop.test.runTest
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.channels.Channel
-import kotlinx.coroutines.channels.ClosedReceiveChannelException
-import kotlinx.coroutines.channels.toList
-import kotlinx.coroutines.launch
-import kotlin.test.Test
-import kotlin.test.assertEquals
-import kotlin.test.assertTrue
-
-class OperatorsTest : AsyncTest() {
-
-
-    @Test
-    fun combineLatestShouldGiveTheCombinedElements() = runTest {
-        val source1 = Channel<Int>()
-        val source2 = Channel<Int>()
-        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
-
-        expect(1)
-        launch(Dispatchers.Unconfined) {
-            expect(2)
-            assertEquals(2 to 3, result.receive())
-            expect(5)
-            assertEquals(2 to 4, result.receive())
-            expect(7)
-            assertEquals(5 to 4, result.receive())
-            expect(9)
-            assertEquals(5 to 6, result.receive())
-            expect(11)
-            assertThrows<ClosedReceiveChannelException> { result.receive() }
-            expect(13)
-        }
-        expect(3)
-        source1.send(1)
-        source1.send(2)
-        expect(4)
-        source2.send(3)
-        expect(6)
-        source2.send(4)
-        expect(8)
-        source1.send(5)
-        source1.close()
-        expect(10)
-        source2.send(6)
-        expect(12)
-        source2.close()
-        finish(14)
-    }
-
-    @Test
-    fun cancellingAChannelMergedByCombineLatestShouldCancelAllTheSources() = runTest {
-        val source1 = Channel<Int>()
-        val source2 = Channel<Int>()
-        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
-
-        result.cancel()
-
-        assertTrue(source1.isClosedForReceive)
-        assertTrue(source2.isClosedForReceive)
-
-        assertThrows<Exception> { source1.receive() }
-        assertThrows<Exception> { source2.receive() }
-    }
-
-    @Test
-    fun ifASourceMergedByCombineLatestFailsTheResultShouldFailAndTheOtherSourcesShouldBeCancelled() = runTest {
-        val source1 = Channel<Int>()
-        val source2 = failedReceiveChannel<Int>(Exception("my exception"))
-        val result = source1.combineLatestWith(source2) { v1, v2 -> v1 to v2 }
-
-        assertTrue(source1.isClosedForReceive, "other source should be closed for receive")
-        assertTrue(result.isClosedForReceive, "result should be closed for receive")
-
-        assertThrows<Exception> { source1.receive() }
-        assertEquals("my exception", assertThrows<Exception> { result.receive() }.message)
-    }
-
-    @Test
-    fun itShouldBePossibleToUseCombineLatestWithFiniteChannel() = runTest {
-        val result = receiveChannelOf(1).combineLatestWith(receiveChannelOf('a')) { v1, v2 -> v1 to v2 }
-
-        assertEquals(1 to 'a', result.receive())
-        assertThrows<ClosedReceiveChannelException> { result.receive() }
-    }
-
-    @Test
-    fun switchMapShouldEmitItemsOfTheNewSourceOnly() = runTest {
-        val sources = (0..2).map { Channel<Char>(2) }
-        val switch = Channel<Int>()
-        val result = switch.switchMap { sources[it] }
-
-        expect(1)
-        launch(Dispatchers.Unconfined) {
-            expect(2)
-            assertEquals('a', result.receive())
-            assertEquals('b', result.receive())
-            expect(5)
-            assertEquals('c', result.receive())
-            assertEquals('d', result.receive())
-            expect(7)
-            assertEquals('y', result.receive())
-            expect(9)
-            assertEquals('e', result.receive())
-            assertEquals('f', result.receive())
-            expect(11)
-            assertThrows<ClosedReceiveChannelException> { result.receive() }
-            expect(14)
-        }
-
-        expect(3)
-        sources[0].send('a')
-        sources[0].send('b')
-        sources[1].send('c')
-        sources[1].send('d')
-        sources[2].send('e')
-        sources[2].send('f')
-        expect(4)
-        switch.send(0) // start the first source
-        expect(6)
-        switch.send(1) // start the second source
-        expect(8)
-        assertThrows<Exception> { sources[0].send('x') } // the first source should have been cancelled
-        sources[1].send('y')
-        expect(10)
-        sources[1].close() // should have no impact
-        switch.send(2) // start the third source
-        expect(12)
-        switch.close() // should have no impact
-        expect(13)
-        sources[2].close() // should close the result as the switch is closed
-        finish(15)
-    }
-
-    @Test
-    fun switchMapOnAnEmptyChannelShouldReturnAnEmptyChannel() = runTest {
-        val result = emptyReceiveChannel<Int>().switchMap { receiveChannelOf(1, 2, 3) }
-
-        assertTrue(result.isClosedForReceive)
-        assertThrows<Exception> { result.receive() }
-    }
-
-    @Test
-    fun cancellingAChannelCreatedByMergeMapShouldCancelTheCurrentSource() = runTest {
-        val source = Channel<Int>()
-        val result = receiveChannelOf(1).switchMap { source }
-
-        result.cancel()
-
-        assertTrue(source.isClosedForReceive)
-        assertThrows<Exception> { source.receive() }
-    }
-
-    @Test
-    fun ifASourceReturnedBySwitchMapFailsTheResultShouldFailWithTheSameException() = runTest {
-        val source = Channel<Int>()
-        val result = receiveChannelOf(1).switchMap { source }
-
-        source.close(Exception("my exception"))
-
-        assertTrue(result.isClosedForReceive)
-        assertEquals("my exception", assertThrows<Exception> { result.receive() }.message)
-    }
-
-    @Test
-    fun distinctUntilChangedShouldNotSendTwiceTheSameValueInARow() = runTest {
-        val receivedValues = receiveChannelOf(1, 2, 2, 2, 3, 2, 1, 1).distinctUntilChanged().toList()
-
-        assertEquals(listOf(1, 2, 3, 2, 1), receivedValues)
-    }
-}
